minirouter: fix interface sorting in findEth()
